Changes to be committed: 	modified:   port/boards/labplus_Ledong_pro/bsp_audio_board.c 	modified:   port/boards/labplus_Ledong_pro/bsp_audio_board.h 	modified:   port/boards/labplus_Ledong_pro/mpconfigboard.cmake 	modified:   port/drivers/audio/audio.c 	deleted:    port/drivers/audio/es8388.c 	deleted:    port/drivers/audio/include/audio_hal.h 	modified:   port/drivers/audio/include/bsp_audio.h 	deleted:    port/drivers/audio/include/es8388.h 	deleted:    port/drivers/audio/include/esxxx_common.h 	modified:   port/drivers/audio/include/player.h 	modified:   port/drivers/audio/include/recorder.h 	modified:   port/drivers/audio/include/wav_codec.h 	modified:   port/drivers/audio/player.c 	modified:   port/drivers/audio/recorder.c 	modified:   port/drivers/audio/wav_codec.c 	modified:   port/test/1.wav 	deleted:    port/test/11.wav 	deleted:    port/test/2.wav 	deleted:    port/test/3.wav 	deleted:    port/test/4.wav

+++ port/drivers/audio/include/bsp_audio.h
@@ -60,7 +60,7 @@ esp_err_t bsp_codec_dev_delete(void);
 esp_err_t bsp_codec_dev_open(int sample_rate, int channel_format, int bits_per_sample);
 esp_err_t bsp_codec_dev_close(void);
 
-esp_err_t bsp_audio_play(const int16_t* data, int length, TickType_t ticks_to_wait);
+esp_err_t bsp_audio_play(const int8_t *data, int length, TickType_t ticks_to_wait);
 
 /**
  * @brief Get the record pcm data.
@@ -73,7 +73,7 @@ esp_err_t bsp_audio_play(const int16_t* data, int length, TickType_t ticks_to_wa
  *    - ESP_OK                  Success
  *    - Others                  Fail
  */
-esp_err_t bsp_get_feed_data(bool is_get_raw_channel, uint8_t *buffer, int buffer_len);
+esp_err_t bsp_get_feed_data(bool is_get_raw_channel, int8_t *buffer, int buffer_len);
 
 int bsp_get_feed_channel(void);
